requirements.txt
langchain
langchain-community
langchain-openai
openai
duckduckgo-search
python-dotenv